# [startup performance] Stop cloning the whole list of DType factor when we don't need to