--- a/numbat/src/typechecker/constraints.rs
+++ b/numbat/src/typechecker/constraints.rs
@@ -1,4 +1,4 @@
-use std::rc::Rc;
+use std::sync::Arc;
 
 use compact_str::{format_compact, CompactString};
 
@@ -308,7 +308,7 @@ impl Constraint {
             Constraint::EqualScalar(d) if d == &DType::scalar() => Some(Satisfied::trivially()),
             Constraint::EqualScalar(dtype) => match dtype.split_first_factor() {
                 Some(((DTypeFactor::TVar(tv), k), rest)) => {
-                    let result = DType::from_factors(Rc::new(
+                    let result = DType::from_factors(Arc::new(
                         rest.iter().map(|(f, j)| (f.clone(), -j / k)).collect(),
                     ));
                     Some(Satisfied::with_substitution(Substitution::single(