@sharkdp 2024-12-27
Ok, let's just take the 13% win for now and revisit this when it comes up again. Thank you very much!